Tested the code. Fixed errors. Moved to uv from poetry
