+++ parser/utils.py
@@ -3,7 +3,7 @@ async def parse_schedule(schedule_string):
     schedule_dict = {}
     parts = schedule_string.split()
 
-    for i in range(len(parts), 2):
+    for i in range(0, len(parts), 2):
         if i + 1 < len(parts):
             day = parts[i].strip(':')
             time_range = parts[i + 1]
