+++ pyproject.toml
@@ -1,22 +1,15 @@
-[tool.poetry]
+[project]
 name = "drugs-availability"
 version = "0.1.0"
-description = ""
-authors = ["Smitona <jarofsmetana@yandex.ru>"]
+description = "Add your description here"
 readme = "README.md"
-package-mode = false
-
-[tool.poetry.dependencies]
-python = "^3.12"
-requests = "^2.32.3"
-python-dotenv = "^1.0.1"
-sqlalchemy = "^2.0.39"
-psycopg2 = "^2.9.10"
-psycopg2-binary = "^2.9.10"
-aiogram = "^3.19.0"
-aiosqlite = "^0.21.0"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires-python = ">=3.12"
+dependencies = [
+    "aiogram>=3.20.0.post0",
+    "aiosqlite>=0.21.0",
+    "psycopg2>=2.9.10",
+    "psycopg2-binary>=2.9.10",
+    "python-dotenv>=1.1.0",
+    "requests>=2.32.3",
+    "sqlalchemy>=2.0.40",
+]
